+++ test/models/animate_to_event_forwarding_test.dart
@@ -8,12 +8,15 @@ import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:kalender/kalender.dart';
 
+import '../utilities.dart';
+
 /// Records what [KalenderController.animateToEvent] passes down.
 class _RecordingViewController extends MultiDayViewController {
   _RecordingViewController({
     required super.viewConfiguration,
     required super.floatingVisibleRange,
     required super.visibleEvents,
+    required super.initial,
   });
 
   Duration? pageDuration;
@@ -48,6 +51,7 @@ void main() {
       ),
       floatingVisibleRange: controller.floatingVisibleRange,
       visibleEvents: controller.visibleEvents,
+      initial: todaySnapshot(),
     );
     controller.attach(viewController);